// frontend/src/openapi/models/Portfolio.ts
/* tslint:disable */
/* eslint-disable */
/**
 * share-po
 * ポートフォリオを共有、評価できるサイトです
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { User } from './User';
import {
    UserFromJSON,
    UserFromJSONTyped,
    UserToJSON,
} from './User';

/**
 * 
 * @export
 * @interface Portfolio
 */
export interface Portfolio {
    /**
     * 
     * @type {number}
     * @memberof Portfolio
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Portfolio
     */
    name?: string;
    /**
     * 
     * @type {User}
     * @memberof Portfolio
     */
    user?: User;
    /**
     * 
     * @type {string}
     * @memberof Portfolio
     */
    description?: string;
    /**
     * 
     * @type {number}
     * @memberof Portfolio
     */
    status?: number;
    /**
     * 
     * @type {Array<number>}
     * @memberof Portfolio
     */
    tags?: Array<number>;
    /**
     * 
     * @type {string}
     * @memberof Portfolio
     */
    githubLink?: string;
    /**
     * 
     * @type {string}
     * @memberof Portfolio
     */
    shareLink?: string;
}

/**
 * Check if a given object implements the Portfolio interface.
 */
export function instanceOfPortfolio(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function PortfolioFromJSON(json: any): Portfolio {
    return PortfolioFromJSONTyped(json, false);
}

export function PortfolioFromJSONTyped(json: any, ignoreDiscriminator: boolean): Portfolio {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'name': !exists(json, 'name') ? undefined : json['name'],
        'user': !exists(json, 'user') ? undefined : UserFromJSON(json['user']),
        'description': !exists(json, 'description') ? undefined : json['description'],
        'status': !exists(json, 'status') ? undefined : json['status'],
        'tags': !exists(json, 'tags') ? undefined : json['tags'],
        'githubLink': !exists(json, 'githubLink') ? undefined : json['githubLink'],
        'shareLink': !exists(json, 'shareLink') ? undefined : json['shareLink'],
    };
}

export function PortfolioToJSON(value?: Portfolio | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
        'user': UserToJSON(value.user),
        'description': value.description,
        'status': value.status,
        'tags': value.tags,
        'githubLink': value.githubLink,
        'shareLink': value.shareLink,
    };
}